core/pyproject.toml
[tool.poetry]
name = "mri-segmentation"
version = "0.1.0"
description = "Patient MRI's AI assisted imaging segmentation"
authors = ["GPDB"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
uvicorn = {extras = ["standard"], version = "^0.13.4"}
Django = "^3.2"
djangorestframework = "^3.12.4"
django-filter = "^2.4.0"
django-js-reverse = "^0.9.1"
django-phonenumber-field = "^5.1.0"
psycopg2 = "^2.8.6"
django-cors-headers = "^3.7.0"
Pillow = "^8.2.0"
pydicom = "^2.1.2"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
flake8 = "^3.8.4"
flake8-bugbear = "^20.11.1"
flake8-comprehensions = "^3.3.1"
flake8-no-types = "^1.1.1"
flake8-tidy-imports = "^4.2.1"
ipdb = "^0.13.3"
ipython = "7.15.0"
isort = "^5.6.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-versions = ["py36"]

[tool.isort]
atomic = true
profile = "black"
sections = ["FUTURE", "STDLIB", "DJANGO", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
known_django = ["django", "rest_framework"]
combine_as_imports = true
line_length = 120
